refactor: cleans up scripts folder
