--- packages/sdk/scripts/builder
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-rm -rf ./dist
-tsc
