--- packages/sdk/scripts/checkRetryableStatus.ts
@@ -1,37 +1,0 @@
-/*
- * Copyright 2021, Offchain Labs, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-/* eslint-env node */
-'use strict'
-
-import { checkRetryableStatus } from './lib'
-
-import args from './getCLargs'
-
-if (!args.txid) {
-  throw new Error('Include txid (--txid 0xmytxid)')
-}
-const txId = args.txid as string
-
-if (!txId) {
-  throw new Error('Need to set an L1 hash')
-}
-
-checkRetryableStatus(txId)
-  .then(() => process.exit(0))
-  .catch(error => {
-    console.error(error)
-    process.exit(1)
-  })
