--- packages/sdk/scripts/setArbCustomGateways.ts
@@ -1,33 +1,0 @@
-/*
- * Copyright 2021, Offchain Labs, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-/* eslint-env node */
-'use strict'
-
-import { setArbCustomGateways } from './lib'
-import args from './getCLargs'
-
-if (!args.address) {
-  throw new Error('Include token address (--address 0xmyaddress)')
-}
-
-const tokens = [args.address as string]
-if (tokens.length === 0) {
-  throw new Error('Include some tokens to set')
-}
-
-setArbCustomGateways(tokens).then(() => {
-  console.log('done')
-})
